Fixed traceback and setup a codeblocks project.
Fixed an off by one error in the traceback code.  I've also created a codeblocks project,
which is probably easier to work with than vim and gdb.  I haven't gotten around to refactoring
the needle function, but I think we could reduce the duplicate code a lot if use callback functions
for parts that change between implementations.